requirements/local.txt
-r base.txt
coverage==3.6
django-discover-runner==0.2.2
django-debug-toolbar==0.9.4
selenium==2.35.0

